Remove file when removing document

+++ java/src/main/java/dev/niziolek/pdfsearch/DocumentService.java
@@ -2,12 +2,15 @@ package dev.niziolek.pdfsearch;
 
 import lombok.AllArgsConstructor;
 import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.gridfs.GridFsTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 @AllArgsConstructor
 @Service
@@ -27,6 +30,9 @@ public class DocumentService {
   }
 
   public void deleteDocument(String id) {
+    Optional<Document> document = documentRepository.findById(id);
+    if (document.isEmpty()) return;
+    gridFsTemplate.delete(new Query().addCriteria(Criteria.where("_id").is(document.get().content)));
     documentRepository.deleteById(id);
   }
